+++ src/main/java/com/rpm/web/magazine/MagazineInit.java
@@ -1,5 +1,6 @@
 package com.rpm.web.magazine;
 
+import com.rpm.web.util.MakeCarDummyList;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
@@ -29,7 +30,6 @@ public class MagazineInit implements ApplicationRunner {
         ArticleCrawler crawler = new ArticleCrawler( articleRepository , extractedWordRepository );
         if (articleRepository.count() == 0 ) crawler.motorgraphComtroller();
         if (extractedWordRepository.count() == 0 ) crawler.extractor();
-        System.out.println("---------- MagazineInit ------------");
         //morphemeAnalyzer();
 
         System.out.println( formattedTime1 + "  INFO 18844 --- [           MagazineInit ]         : MagazineInit End ");
