+++ vue/public/index.html
@@ -16,11 +16,7 @@
   <link rel="stylesheet" type="text/css"  href="css/plugin/selectric.css">
   <link rel="stylesheet" type="text/css" href="css/re_layout.css">
 
-
 <body>
-<noscript>
-  <strong>We're sorry but vue doesn't work properly without JavaScript enabled. Please enable it to continue.</strong>
-</noscript>
 
 <div id="app"></div>
 <!-- built files will be auto injected -->
