+++ vue/src/RPM.vue
@@ -21,8 +21,6 @@
                                 <router-link to="/magazine">magazine </router-link>
                                 <router-link to="/chartMaker">chartMaker </router-link>
                                 <router-link to="/sns">SNS(REVIEW) </router-link>
-                                <router-link to="/mapList">가까운 주유소 찾기 </router-link>
-
                             </Slide>
                         </div>
                         <div v-if="this.auth">
@@ -36,7 +34,6 @@
                                 <router-link to="/sns">SNS(REVIEW) </router-link>
                                 <router-link to="/magazine">magazine </router-link>
                                 <router-link to="/chartMaker">chartMaker </router-link>
-                                <router-link to="/mapList">가까운 주유소 찾기 </router-link>
                             </Slide >
                         </div>
                     </li>
