+++ vue/src/components/magazine/DatacenterChart.vue
