+++ vue/src/components/magazine/MagazineArticle.vue
