+++ vue/src/components/magazine/MagazineCarList.vue
