+++ vue/src/components/magazine/chartjs/baseCharts.js
